legacy: OAIHarvest removal
* INCOMPATIBLE Removes legacy OAIHarvest module. (addresses #3233)

* Moves required parts from legacy OAIHarvest to oaiharvester module
  in preparation for new `invenio-oaiharvester` module.
