+++ invenio/modules/oaiharvester/testsuite/test_utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2014 CERN.
+# Copyright (C) 2014, 2015 CERN.
 #
 # Invenio is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -22,7 +22,7 @@
 import os
 import tempfile
 
-from invenio.testsuite import make_test_suite, run_test_suite, InvenioTestCase
+from invenio.testsuite import InvenioTestCase, make_test_suite, run_test_suite
 
 
 class OAIHarvesterUtils(InvenioTestCase):
@@ -193,6 +193,25 @@ class OAIHarvesterUtils(InvenioTestCase):
 
         self.assertEqual(len(record_extraction_from_file(path_tmp)), 1)
 
+    def test_identifier_filter(self):
+        """oaiharvest - testing identifier filter."""
+        from invenio.modules.oaiharvester.utils import get_identifier_names
+        self.assertEqual(get_identifier_names("oai:mysite.com:1234"),
+                         ["oai:mysite.com:1234"])
+        self.assertEqual(get_identifier_names("oai:mysite.com:1234, oai:example.com:2134"),
+                         ["oai:mysite.com:1234", "oai:example.com:2134"])
+        self.assertEqual(get_identifier_names("oai:mysite.com:1234/testing, oai:example.com:record/1234"),
+                         ["oai:mysite.com:1234/testing", "oai:example.com:record/1234"])
+
+    def test_identifier_filter_special_arXiv(self):
+        """oaiharvest - testing identifier filter for arXiv."""
+        from invenio.modules.oaiharvester.utils import get_identifier_names
+        self.assertEqual(get_identifier_names("oai:arxiv.org:1234.1245"),
+                         ["oai:arXiv.org:1234.1245"])
+        self.assertEqual(get_identifier_names("oai:arXiv.org:1234.1245, arXiv:1234.1245"),
+                         ["oai:arXiv.org:1234.1245", "oai:arXiv.org:1234.1245"])
+        self.assertEqual(get_identifier_names("oai:arXiv.org:1234.12452"),
+                         ["oai:arXiv.org:1234.12452"])
 
 TEST_SUITE = make_test_suite(OAIHarvesterUtils)
