+++ setup.py
@@ -298,7 +298,6 @@ setup(
             'docextract = invenio.legacy.docextract.scripts.docextract:main',
             'gotoadmin = invenio.modules.redirector.scripts.redirector:main',
             'inveniogc = invenio.legacy.websession.scripts.inveniogc:main',
-            'oaiharvest = invenio.legacy.oaiharvest.scripts.oaiharvest:main',
             'oairepositoryupdater = '
             ' invenio.legacy.oairepository.scripts.oairepositoryupdater:main',
             'refextract = invenio.legacy.refextract.scripts.refextract:main',
